requirements.txt
asgiref==3.7.2; python_version >= '3.7'
django>=4.2,<4.3
djangorestframework>=3.14,<3.15
pytz==2023.3
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.7.1; python_version < '3.11'
djangorestframework-simplejwt>=5.2.2,<5.3
drf-spectacular>=0.26.4,<0.27
django-cors-headers>=4.2.0,<4.3